# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/v1/admin/galeri/{galeri}": {
            "get": {
                "operationId": "admin.galeri.show",
                "summary": "Display the specified resource (public)",
                "tags": [
                    "Galeri"
                ],
                "parameters": [
                    {
                        "name": "galeri",
                        "in": "path",
                        "required": true,
                        "description": "The galeri ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/GaleriResource"
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "delete": {
                "operationId": "admin.galeri.destroy",
                "summary": "Remove the specified resource from storage (admin)",
                "tags": [
                    "Galeri"
                ],
                "parameters": [
                    {
                        "name": "galeri",
                        "in": "path",
                        "required": true,
                        "description": "The galeri ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Galeri deleted successfully"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/admin/paket/{paket}": {
            "get": {
                "operationId": "admin.paket.show",
                "summary": "Display the specified resource (public)",
                "tags": [
                    "Paket"
                ],
                "parameters": [
                    {
                        "name": "paket",
                        "in": "path",
                        "required": true,
                        "description": "The paket ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/PaketResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "images"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "admin.paket.update",
                "summary": "Update the specified resource in storage (admin)",
                "tags": [
                    "Paket"
                ],
                "parameters": [
                    {
                        "name": "paket",
                        "in": "path",
                        "required": true,
                        "description": "The paket ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaketUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Paket updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PaketResource"
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "admin.paket.destroy",
                "summary": "Remove the specified resource from storage (admin)",
                "tags": [
                    "Paket"
                ],
                "parameters": [
                    {
                        "name": "paket",
                        "in": "path",
                        "required": true,
                        "description": "The paket ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Paket deleted successfully"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "GaleriResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nama_acara": {
                        "type": "string"
                    },
                    "deskripsi_acara": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "gambar_acara": {
                        "type": "string"
                    },
                    "tanggal_acara": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "nama_acara",
                    "deskripsi_acara",
                    "gambar_acara",
                    "tanggal_acara",
                    "created_at",
                    "updated_at"
                ],
                "title": "GaleriResource"
            },
...
            "KategoriAcaraEnum": {
                "type": "string",
                "enum": [
                    "Pernikahan",
                    "Kantor",
                    "Ulang Tahun",
                    "Arisan",
                    "Umum"
                ],
                "title": "KategoriAcaraEnum"
            },
...
            "PaketKategoriEnum": {
                "type": "string",
                "enum": [
                    "Nasi Box",
                    "Prasmanan",
                    "Snack",
                    "Tumpeng"
                ],
                "title": "PaketKategoriEnum"
            },
            "PaketResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nama_paket": {
                        "type": "string"
                    },
                    "kategori_paket": {
                        "type": "string"
                    },
                    "kategori_acara": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "menu_utama": {
                        "type": "array",
                        "items": {}
                    },
                    "menu_tambahan": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {}
                    },
                    "fasilitas_termasuk": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {}
                    },
                    "catatan_alergen": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "jenis_kemasan": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "min_order": {
                        "type": "integer"
                    },
                    "harga_per_porsi": {
                        "type": "string"
                    },
                    "kapasitas_produksi": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "deskripsi": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "thumbnail": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "images": {
                        "type": "array",
                        "items": {}
                    },
                    "is_best_seller": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "nama_paket",
                    "kategori_paket",
                    "kategori_acara",
                    "menu_utama",
                    "menu_tambahan",
                    "fasilitas_termasuk",
                    "catatan_alergen",
                    "jenis_kemasan",
                    "min_order",
                    "harga_per_porsi",
                    "kapasitas_produksi",
                    "deskripsi",
                    "thumbnail",
                    "is_best_seller",
                    "created_at",
                    "updated_at"
                ],
                "title": "PaketResource"
            },
...
            "PaketUpdateRequest": {
                "type": "object",
                "properties": {
                    "nama_paket": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "kategori_paket": {
                        "$ref": "#/components/schemas/PaketKategoriEnum"
                    },
                    "kategori_acara": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/KategoriAcaraEnum"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "menu_utama": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "maxLength": 255
                        },
                        "minItems": 1
                    },
                    "menu_tambahan": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    "fasilitas_termasuk": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    "catatan_alergen": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "jenis_kemasan": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    },
                    "min_order": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 1
                    },
                    "harga_per_porsi": {
                        "type": "number",
                        "minimum": 0
                    },
                    "kapasitas_produksi": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 1
                    },
                    "deskripsi": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "thumbnail": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 2048
                    },
                    "is_best_seller": {
                        "type": [
                            "boolean",
                            "null"
                        ]
                    }
                },
                "title": "PaketUpdateRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}